Backend/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Django==3.2.10
django-cors-headers==3.10.1
django-jalali-date==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
importlib-metadata==4.10.1
jdatetime==4.0.0
mypy-extensions==0.4.3
mysqlclient
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.4.1
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tomli==1.2.3
typed-ast==1.5.2
typing_extensions==4.0.1
urllib3==1.26.8
zipp==3.7.0
